pyproject.toml
[project]
name = "proyecto-supermercado"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "mysql-connector-python>=8.0",
    "werkzeug>=3.0.0",
    "python-dotenv>=1.0.0"
]
